Update index type to u32, rerun examples

+++ python/uatk_spc/builder.py
@@ -93,7 +93,7 @@ class Builder(Reader):
                     self.time_use_diaries,
                     pl.int_range(0, self.time_use_diaries.shape[0], eager=True)
                     .rename("index")
-                    .cast(pl.UInt64)
+                    .cast(pl.UInt32)
                     .to_frame(),
                 ],
                 how="horizontal",
